+++ tests/numbers_windows_test.go
@@ -1,0 +1,9 @@
+package tests
+
+import (
+	"testing"
+)
+
+func TestDifferentNumberBasesSuccess(t *testing.T) {
+	testDifferentNumberBasesSuccess(t, transpileBatch)
+}